RAG/requirement.txt.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
groq
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
